+++ library/encryptDecrypt.h
@@ -45,9 +45,13 @@ information (HMAC and IV) to allow for decryption.
 /* Function Introduction */
 
 void encryptHMAC(uint8_t*, uint16_t*, uint8_t*, uint8_t*, uint16_t);
+	/* Encrypts with AES, and appends HMAC (in a specific way) */
 void decryptHMAC(uint8_t*, uint16_t*, uint8_t*, uint8_t*, uint16_t);
-void encrypt(uint8_t*, uint8_t*, uint8_t*, uint8_t*, uint16_t);
-void decrypt(uint8_t*, uint8_t*, uint8_t*, uint8_t*, uint16_t);
+	/* Decrypts with AES, and appends HMAC (in a specific way) */
+void encrypt(uint8_t*, uint16_t*, uint8_t*, uint8_t*, uint16_t);
+	/* Encrypts with AES */
+void decrypt(uint8_t*, uint16_t*, uint8_t*, uint8_t*, uint16_t);
+	/* Decrypts with AES */
 
 #endif
